Comprehensive bug fixes from code review (round 3)
Critical/High:
- http.cpp: fix page_catf truncation corrupting HTML — clamp vsnprintf return
  so page_pos advances on overflow instead of leaving partial bytes to be
  overwritten by the next call.
- cue.cpp: lift elapsed_ms to uint64_t (with int64_t effective_start) to
  eliminate the ~49-day wrap during long playback.
- wasm_wrapper.cpp: factor cleanup into wasm_cleanup_runtime() and call it
  on the m3_RunStart error path so file slots, string pool, low heap, and
  current-path state don't leak into the next run.
- conez_wifi.cpp: s_tx_bytes / s_rx_bytes now std::atomic<uint32_t> with
  fetch_add(relaxed) so concurrent LWIP callbacks across cores don't lose
  increments or tear reads.

Medium:
- c2wasm cw_malloc/realloc/calloc: abort on NULL via cw_fatal (embedded) or
  exit(1) (standalone) — every call site is safe without per-site checks.
- c2wasm buf.c: guard nc *= 2 with INT_MAX/2 to prevent capacity overflow.
- wasm_imports_file.cpp: validate fseek target against file size via ftell +
  fstat; reject out-of-range offsets.
- psram.cpp: SPI2_WAIT(cond) macro replaces all 8 unbounded
  while(GPSPI2.cmd.X) spins with a 1M-iter cap (~4ms @ 240MHz).
- telnet.cpp/.h: new iac_sb_bytes counter aborts subnegotiation after 256
  bytes so a malformed IAC stream can't pin a client in state 3.
- nmea.cpp: RMC/GGA commit only overwrites lat/lon when s_has_fix is true,
  so DR/simulated fixes don't clobber last-good position; location_valid
  still reflects fix status.
- gps.cpp datetime_to_epoch_ms: range-check year/month/day/hour/minute/second
  and return 0 on bad input.
- config.cpp: new config_mutex serializes config_set_field and config_save
  so cross-task readers don't observe torn string-field writes.
- editor.cpp: write to <path>.tmp then rename — crash mid-save leaves the
  original file untouched.

Low:
- led.cpp: led_setup() frees old CRGB arrays on re-entry; rmt_init guarded
  by static bool so RMT channels/encoder can't double-create.
- led.cpp: led_mutex held across the whole led_push_hw so writers can't race
  with the render task on grb_buf or the CRGB buffers.
- wasm_imports_string.cpp: pool_free and low_heap_free coalesce adjacent
  free blocks, shrink bump when the coalesced block is at the top, and
  prune trailing dead records.
- printManager.cpp: ANSI stripper handles ESC] OSC (BEL/ST-terminated) and
  ESC X single-char sequences, not just ESC [ CSI.
- effects.cpp/.h: per-effect statics moved to module scope with s_* prefix;
  new effects_reset_all() zeroes them when switching effects.

New feature:
- http.cpp: /nvs handler now dumps NVS contents grouped by namespace with
  typed values, escaped strings, and blob hex dumps (was "FIXME..." stub).

+++ firmware/src/cue/cue.cpp
@@ -172,8 +172,9 @@ void cue_loop(void)
     uint64_t now_ms = get_epoch_ms();
     if (now_ms == 0) { xSemaphoreGive(cue_mutex); return; }
 
-    // Simple subtraction — epoch time never wraps at midnight
-    uint32_t elapsed_ms = (now_ms > music_start_ms) ? (uint32_t)(now_ms - music_start_ms) : 0;
+    // Simple subtraction — epoch time never wraps at midnight.
+    // 64-bit elapsed avoids the ~49-day wrap of uint32_t.
+    uint64_t elapsed_ms = (now_ms > music_start_ms) ? (now_ms - music_start_ms) : 0;
 
     // Walk cue list from cursor forward
     while (cue_cursor < cue_count) {
@@ -181,11 +182,11 @@ void cue_loop(void)
 
         // Compute this cone's effective start time including spatial offset
         int32_t spatial_off = compute_spatial_offset(cue);
-        int32_t effective_start = (int32_t)cue->start_ms + spatial_off;
+        int64_t effective_start = (int64_t)cue->start_ms + spatial_off;
         if (effective_start < 0) effective_start = 0;
 
         // Not yet time for this cue?
-        if ((uint32_t)effective_start > elapsed_ms)
+        if ((uint64_t)effective_start > elapsed_ms)
             break;
 
         // Check group targeting
